requirements.txt
numpy>=1.19.5
pandas==1.1.5
requests==2.13.0
aiohttp==3.7.3
websockets==8.1
python-telegram-bot==13.1
cbpro==1.1.4